Size is an int, not a float!

+++ plug-ins/common/edu.illinois.codingtracker.tests/src/edu/oregonstate/codingtracker/tests/recommender/ItemSet.java
@@ -42,7 +42,7 @@ public class ItemSet implements Iterable<Item> {
 		return items.contains(item);
 	}
 
-	public float size() {
+	public int size() {
 		return size;
 	}
